+++ lib/core/common_widget/adaptive_overflow_text.dart
