+++ lib/feature/definition/presentation/component/definition_tile.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:like_button/like_button.dart';
 
-import '../../../../common_widget/adaptive_overflow_text.dart';
+import '../../../../core/common_widget/adaptive_overflow_text.dart';
 import '../../../../util/constant/color_scheme.dart';
 import '../../../../util/extension/date_time_extension.dart';
 import '../../application/definition_service.dart';
