+++ lib/feature/definition/repository/definition_repository.dart
@@ -3,8 +3,8 @@ import 'dart:math';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
+import '../../../core/common_provider/firebase_provider.dart';
 import '../../../util/constant/config_constant.dart';
-import '../../../util/firebase_provider.dart';
 import '../domain/definition_id_list_state.dart';
 import 'entity/definition_document.dart';
