+++ lib/feature/word/repository/word_repository.dart
@@ -1,7 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../../../util/firebase_provider.dart';
+import '../../../core/common_provider/firebase_provider.dart';
 import 'entity/word_document.dart';
 
 part 'word_repository.g.dart';
