+++ test/core/common_provider/is_loading_overlay_state_test.mocks.dart
@@ -1,0 +1,40 @@
+// Mocks generated by Mockito 5.4.2 from annotations
+// in teigi_app/test/core/common_provider/is_loading_overlay_state_test.dart.
+// Do not manually edit this file.
+
+// ignore_for_file: no_leading_underscores_for_library_prefixes
+import 'package:mockito/mockito.dart' as _i1;
+
+import 'is_loading_overlay_state_test.dart' as _i2;
+
+// ignore_for_file: type=lint
+// ignore_for_file: avoid_redundant_argument_values
+// ignore_for_file: avoid_setters_without_getters
+// ignore_for_file: comment_references
+// ignore_for_file: implementation_imports
+// ignore_for_file: invalid_use_of_visible_for_testing_member
+// ignore_for_file: prefer_const_constructors
+// ignore_for_file: unnecessary_parenthesis
+// ignore_for_file: camel_case_types
+// ignore_for_file: subtype_of_sealed_class
+
+/// A class which mocks [Listener].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockListener extends _i1.Mock implements _i2.Listener<bool> {
+  @override
+  void call(
+    bool? previous,
+    bool? next,
+  ) =>
+      super.noSuchMethod(
+        Invocation.method(
+          #call,
+          [
+            previous,
+            next,
+          ],
+        ),
+        returnValueForMissingStub: null,
+      );
+}
